refactor(chat-loop): split background executors
- extract model loading into executors/model_loader.rs
- move MCP init flow into executors/mcp_init.rs with focused tests
- move MCP tool, prompt, sampling, refresh, and error flows to
  executors/mcp_tools.rs
- add ExecutorContext to share app handle, dispatcher, cancel token,
  and terminal size
- keep event_loop.rs focused on polling and dispatch, and document
  chat_loop/executors in README
