+++ src/ui/chat_loop/event_loop.rs
@@ -18,9 +18,6 @@ use std::{
     time::{Duration, Instant},
 };
 
-// Default budget for MCP sampling when the server does not provide a timeout hint.
-const MCP_SAMPLING_DEFAULT_TIMEOUT: Duration = Duration::from_secs(300);
-const MCP_SAMPLING_SEND_TIMEOUT: Duration = Duration::from_secs(10);
 const ACTIVE_POLL_INTERVAL_MS: u64 = 10;
 const ANIMATION_POLL_INTERVAL_MS: u64 = 16;
 const IDLE_POLL_INTERVAL_MS: u64 = 100;
@@ -28,30 +25,24 @@ const IDLE_SLEEP_MS: u64 = 50;
 const INPUT_BURST_POLL_INTERVAL_MS: u64 = 1;
 const INPUT_BURST_WINDOW_MS: u64 = 200;
 
-use ratatui::crossterm::event::{self, Event, KeyEventKind, KeyModifiers};
-use ratatui::prelude::Size;
-use serde::Serialize;
-use serde_json;
-use tokio::sync::mpsc;
-use tokio_util::sync::CancellationToken;
-use tracing::debug;
-
-use crate::api::models::fetch_models;
-use crate::api::ChatRequest;
 use crate::character::CharacterService;
 use crate::core::app::{
     apply_actions, AppActionContext, AppActionDispatcher, AppActionEnvelope, AppCommand,
-    ComposeAction, InspectAction, InspectMode, ModelPickerRequest, PickerAction, StatusAction,
-    StreamingAction,
+    ComposeAction, InspectAction, InspectMode, StreamingAction,
 };
-use crate::core::chat_stream::{request_chat_completion, ChatStreamService, StreamMessage};
-use crate::core::mcp_auth::McpTokenStore;
-use crate::core::mcp_sampling::map_finish_reason;
-use crate::core::message::AppMessageKind;
+use crate::core::chat_stream::{ChatStreamService, StreamMessage};
 use crate::ui::renderer::ui;
-use rust_mcp_schema::schema_utils::ResultFromClient;
-use rust_mcp_schema::{CreateMessageContent, CreateMessageResult, Role, TextContent};
+use ratatui::crossterm::event::{self, Event, KeyEventKind, KeyModifiers};
+use ratatui::prelude::Size;
+use tokio::sync::mpsc;
 
+use super::executors::mcp_init::spawn_mcp_initializer;
+use super::executors::mcp_tools::{
+    spawn_mcp_prompt_call, spawn_mcp_refresh, spawn_mcp_sampling_call, spawn_mcp_server_error,
+    spawn_mcp_tool_call,
+};
+use super::executors::model_loader::spawn_model_picker_loader;
+use super::executors::ExecutorContext;
 use super::keybindings::{build_mode_aware_registry, KeyContext, KeyResult, ModeAwareRegistry};
 use super::lifecycle::{
     apply_cursor_color_to_terminal, restore_terminal, setup_terminal, SharedTerminal,
@@ -62,644 +53,6 @@ use super::AppHandle;
 #[derive(Debug)]
 pub enum UiEvent {
     Crossterm(Event),
-}
-
-fn spawn_model_picker_loader(dispatcher: AppActionDispatcher, request: ModelPickerRequest) {
-    tokio::spawn(async move {
-        let ModelPickerRequest {
-            client,
-            base_url,
-            api_key,
-            provider_name,
-            default_model_for_provider,
-        } = request;
-
-        let fetch_result = fetch_models(&client, &base_url, &api_key, &provider_name)
-            .await
-            .map_err(|e| e.to_string());
-
-        let action = match fetch_result {
-            Ok(models_response) => PickerAction::ModelPickerLoaded {
-                default_model_for_provider,
-                models_response,
-            },
-            Err(e) => PickerAction::ModelPickerLoadFailed { error: e },
-        };
-
-        dispatcher.dispatch_many([action], AppActionContext::default());
-    });
-}
-
-fn spawn_mcp_initializer(
-    app: AppHandle,
-    dispatcher: AppActionDispatcher,
-    request_tx: mpsc::UnboundedSender<crate::mcp::events::McpServerRequest>,
-) {
-    tokio::spawn(async move {
-        let mcp_disabled = app.read(|app| app.session.mcp_disabled).await;
-        if mcp_disabled {
-            dispatcher.dispatch_many(
-                [StreamingAction::McpInitCompleted],
-                AppActionContext::default(),
-            );
-            return;
-        }
-
-        let has_enabled_servers = app
-            .read(|app| app.mcp.servers().any(|server| server.config.is_enabled()))
-            .await;
-
-        if !has_enabled_servers {
-            dispatcher.dispatch_many(
-                [StreamingAction::McpInitCompleted],
-                AppActionContext::default(),
-            );
-            return;
-        }
-
-        let keyring_enabled = app.read(|app| !app.session.startup_env_only).await;
-        let token_store = McpTokenStore::new_with_keyring(keyring_enabled);
-
-        let config = app.read(|app| app.config.clone()).await;
-        let mut mcp = crate::mcp::client::McpClientManager::from_config(&config);
-        mcp.set_request_sender(request_tx.clone());
-        mcp.connect_all(&token_store).await;
-
-        let server_ids: Vec<String> = mcp
-            .servers()
-            .filter(|server| server.config.is_enabled())
-            .map(|server| server.config.id.clone())
-            .collect();
-
-        for server_id in server_ids {
-            mcp.refresh_server_metadata_concurrently(&server_id).await;
-        }
-
-        app.update(|app| {
-            app.mcp = mcp;
-            app.session.mcp_init_in_progress = false;
-            app.session.mcp_init_complete = true;
-        })
-        .await;
-
-        dispatcher.dispatch_many(
-            [StreamingAction::McpInitCompleted],
-            AppActionContext::default(),
-        );
-    });
-}
-
-fn spawn_mcp_tool_call(
-    app: AppHandle,
-    dispatcher: AppActionDispatcher,
-    request: crate::core::app::session::ToolCallRequest,
-) {
-    tokio::spawn(async move {
-        let cancel_token = app
-            .read(|app| app.session.stream_cancel_token.clone())
-            .await;
-        let term_size = app.read(|app| app.ui.last_term_size).await;
-        let ctx = AppActionContext {
-            term_width: term_size.width,
-            term_height: term_size.height,
-        };
-
-        let mut call_context = match app
-            .read(|app| app.mcp.tool_call_context(&request.server_id))
-            .await
-        {
-            Some(context) => context,
-            None => {
-                dispatcher.dispatch_many(
-                    [StreamingAction::ToolCallCompleted {
-                        tool_name: request.tool_name.clone(),
-                        tool_call_id: request.tool_call_id.clone(),
-                        result: Err("MCP server not available.".to_string()),
-                    }],
-                    ctx,
-                );
-                return;
-            }
-        };
-
-        let result = if request
-            .tool_name
-            .eq_ignore_ascii_case(crate::mcp::MCP_READ_RESOURCE_TOOL)
-        {
-            let uri = match request
-                .arguments
-                .as_ref()
-                .and_then(|args| args.get("uri"))
-                .and_then(|value| value.as_str())
-            {
-                Some(uri) => uri.to_string(),
-                None => {
-                    dispatcher.dispatch_many(
-                        [StreamingAction::ToolCallCompleted {
-                            tool_name: request.tool_name.clone(),
-                            tool_call_id: request.tool_call_id.clone(),
-                            result: Err("Resource read requires uri.".to_string()),
-                        }],
-                        ctx,
-                    );
-                    return;
-                }
-            };
-
-            run_cancellable(
-                cancel_token.as_ref(),
-                crate::mcp::client::execute_resource_read(&mut call_context, &uri),
-            )
-            .await
-            .map(|result| serialize_mcp_result(&result))
-        } else if request
-            .tool_name
-            .eq_ignore_ascii_case(crate::mcp::MCP_LIST_RESOURCES_TOOL)
-        {
-            let Some(arguments) = request.arguments.as_ref() else {
-                dispatcher.dispatch_many(
-                    [StreamingAction::ToolCallCompleted {
-                        tool_name: request.tool_name.clone(),
-                        tool_call_id: request.tool_call_id.clone(),
-                        result: Err("Resource list arguments are required.".to_string()),
-                    }],
-                    ctx,
-                );
-                return;
-            };
-
-            let (kind, cursor) =
-                match crate::core::app::actions::parse_resource_list_kind(arguments) {
-                    Ok(values) => values,
-                    Err(err) => {
-                        dispatcher.dispatch_many(
-                            [StreamingAction::ToolCallCompleted {
-                                tool_name: request.tool_name.clone(),
-                                tool_call_id: request.tool_call_id.clone(),
-                                result: Err(err),
-                            }],
-                            ctx,
-                        );
-                        return;
-                    }
-                };
-
-            match kind {
-                crate::core::app::actions::ResourceListKind::Resources => run_cancellable(
-                    cancel_token.as_ref(),
-                    crate::mcp::client::execute_resource_list(&mut call_context, cursor),
-                )
-                .await
-                .map(|result| serialize_mcp_result(&result)),
-                crate::core::app::actions::ResourceListKind::Templates => run_cancellable(
-                    cancel_token.as_ref(),
-                    crate::mcp::client::execute_resource_template_list(&mut call_context, cursor),
-                )
-                .await
-                .map(|result| serialize_mcp_result(&result)),
-            }
-        } else {
-            run_cancellable(
-                cancel_token.as_ref(),
-                crate::mcp::client::execute_tool_call(&mut call_context, &request),
-            )
-            .await
-            .map(|result| serialize_mcp_result(&result))
-        };
-
-        let session_id = call_context.session_id.clone();
-        let error = result.as_ref().err().cloned();
-        app.update(|app| {
-            app.mcp
-                .update_tool_call_session(&call_context.server_id, session_id, error);
-        })
-        .await;
-
-        dispatcher.dispatch_many(
-            [StreamingAction::ToolCallCompleted {
-                tool_name: request.tool_name.clone(),
-                tool_call_id: request.tool_call_id.clone(),
-                result,
-            }],
-            ctx,
-        );
-    });
-}
-
-fn serialize_mcp_result<T: Serialize>(result: &T) -> String {
-    serde_json::to_string_pretty(result)
-        .unwrap_or_else(|_| "Unable to serialize MCP result.".to_string())
-}
-
-async fn run_cancellable<F, T>(
-    cancel_token: Option<&CancellationToken>,
-    operation: F,
-) -> Result<T, String>
-where
-    F: std::future::Future<Output = Result<T, String>>,
-{
-    if let Some(token) = cancel_token {
-        tokio::select! {
-            _ = token.cancelled() => Err("MCP operation interrupted by user.".to_string()),
-            result = operation => result,
-        }
-    } else {
-        operation.await
-    }
-}
-
-fn spawn_mcp_prompt_call(
-    app: AppHandle,
-    dispatcher: AppActionDispatcher,
-    request: crate::core::app::session::McpPromptRequest,
-) {
-    tokio::spawn(async move {
-        let term_size = app.read(|app| app.ui.last_term_size).await;
-        let ctx = AppActionContext {
-            term_width: term_size.width,
-            term_height: term_size.height,
-        };
-
-        let mut call_context = match app
-            .read(|app| app.mcp.prompt_call_context(&request.server_id))
-            .await
-        {
-            Some(context) => context,
-            None => {
-                dispatcher.dispatch_many(
-                    [StreamingAction::McpPromptCompleted {
-                        request,
-                        result: Err("MCP server not available.".to_string()),
-                    }],
-                    ctx,
-                );
-                return;
-            }
-        };
-
-        let result = crate::mcp::client::execute_prompt(&mut call_context, &request).await;
-
-        let session_id = call_context.session_id.clone();
-        let error = result.as_ref().err().cloned();
-        app.update(|app| {
-            app.mcp
-                .update_prompt_session(&call_context.server_id, session_id, error);
-        })
-        .await;
-
-        dispatcher.dispatch_many(
-            [StreamingAction::McpPromptCompleted { request, result }],
-            ctx,
-        );
-    });
-}
-
-fn spawn_mcp_sampling_call(
-    app: AppHandle,
-    dispatcher: AppActionDispatcher,
-    request: crate::core::app::session::McpSamplingRequest,
-) {
-    tokio::spawn(async move {
-        let start = std::time::Instant::now();
-        let cancel_token = app
-            .read(|app| app.session.stream_cancel_token.clone())
-            .await;
-        let term_size = app.read(|app| app.ui.last_term_size).await;
-        let ctx = AppActionContext {
-            term_width: term_size.width,
-            term_height: term_size.height,
-        };
-
-        let (client, base_url, api_key, provider_name, model) = app
-            .read(|app| {
-                (
-                    app.session.client.clone(),
-                    app.session.base_url.clone(),
-                    app.session.api_key.clone(),
-                    app.session.provider_name.clone(),
-                    app.session.model.clone(),
-                )
-            })
-            .await;
-
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request.request.id,
-            "Starting MCP sampling request"
-        );
-
-        let messages = request.messages;
-
-        let stop = if request.request.params.stop_sequences.is_empty() {
-            None
-        } else {
-            Some(request.request.params.stop_sequences.clone())
-        };
-
-        let chat_request = ChatRequest {
-            model: model.clone(),
-            messages,
-            stream: false,
-            tools: None,
-            max_tokens: Some(request.request.params.max_tokens),
-            temperature: request.request.params.temperature,
-            stop,
-        };
-
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request.request.id,
-            "Fetching MCP sampling response context"
-        );
-        let mut context = match app
-            .read(|app| app.mcp.server_request_context(&request.server_id))
-            .await
-        {
-            Some(context) => {
-                debug!(
-                    server_id = %request.server_id,
-                    request_id = ?request.request.id,
-                    "MCP sampling context acquired"
-                );
-                context
-            }
-            None => {
-                debug!(
-                    server_id = %request.server_id,
-                    request_id = ?request.request.id,
-                    elapsed_ms = start.elapsed().as_millis(),
-                    "MCP sampling context missing; aborting response"
-                );
-                dispatcher.dispatch_many([StreamingAction::McpSamplingFinished], ctx);
-                return;
-            }
-        };
-
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request.request.id,
-            provider = %provider_name,
-            model = %model,
-            max_tokens = request.request.params.max_tokens,
-            temperature = ?request.request.params.temperature,
-            "Dispatching MCP sampling completion"
-        );
-        let default_sampling_timeout = MCP_SAMPLING_DEFAULT_TIMEOUT;
-        let sampling_timeout =
-            crate::core::mcp_sampling::sampling_timeout_for_request(&request.request)
-                .unwrap_or(default_sampling_timeout);
-        if sampling_timeout != default_sampling_timeout {
-            debug!(
-                server_id = %request.server_id,
-                request_id = ?request.request.id,
-                timeout_secs = sampling_timeout.as_secs(),
-                "Using MCP sampling timeout hint"
-            );
-        }
-        let completion = match tokio::time::timeout(
-            sampling_timeout,
-            request_chat_completion(&client, &base_url, &api_key, &provider_name, chat_request),
-        )
-        .await
-        {
-            Ok(result) => result,
-            Err(_) => {
-                debug!(
-                    server_id = %request.server_id,
-                    request_id = ?request.request.id,
-                    elapsed_ms = start.elapsed().as_millis(),
-                    timeout_secs = sampling_timeout.as_secs(),
-                    "MCP sampling completion timed out"
-                );
-                Err(format!(
-                    "Sampling timed out after {}s",
-                    sampling_timeout.as_secs()
-                ))
-            }
-        };
-
-        if cancel_token
-            .as_ref()
-            .is_some_and(tokio_util::sync::CancellationToken::is_cancelled)
-        {
-            let session_id = context.session_id.clone();
-            app.update(|app| {
-                app.mcp.update_server_request_session(
-                    &context.server_id,
-                    session_id,
-                    Some("MCP operation interrupted by user.".to_string()),
-                );
-            })
-            .await;
-            dispatcher.dispatch_many([StreamingAction::McpSamplingFinished], ctx);
-            return;
-        }
-
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request.request.id,
-            elapsed_ms = start.elapsed().as_millis(),
-            "Sending MCP sampling response"
-        );
-        let send_timeout = MCP_SAMPLING_SEND_TIMEOUT;
-        let send_operation = tokio::time::timeout(send_timeout, async {
-            match completion {
-                Ok(result) => {
-                    let preview: String = result.content.chars().take(300).collect();
-                    debug!(
-                        server_id = %request.server_id,
-                        request_id = ?request.request.id,
-                        finish_reason = ?result.finish_reason,
-                        elapsed_ms = start.elapsed().as_millis(),
-                        content_chars = result.content.len(),
-                        content_preview = %preview,
-                        "MCP sampling completion received"
-                    );
-                    let content =
-                        CreateMessageContent::from(TextContent::new(result.content, None, None));
-                    let create_result = CreateMessageResult {
-                        content,
-                        meta: None,
-                        model: model.clone(),
-                        role: Role::Assistant,
-                        stop_reason: map_finish_reason(result.finish_reason),
-                    };
-                    crate::mcp::client::send_client_result(
-                        &mut context,
-                        request.request.id.clone(),
-                        ResultFromClient::CreateMessageResult(create_result),
-                    )
-                    .await
-                }
-                Err(err) => {
-                    debug!(
-                        server_id = %request.server_id,
-                        request_id = ?request.request.id,
-                        error = %err,
-                        elapsed_ms = start.elapsed().as_millis(),
-                        "MCP sampling completion failed"
-                    );
-                    crate::mcp::client::send_client_error(
-                        &mut context,
-                        request.request.id.clone(),
-                        rust_mcp_schema::RpcError::internal_error().with_message(&err),
-                    )
-                    .await
-                }
-            }
-        });
-        let send_result = if let Some(token) = cancel_token.as_ref() {
-            tokio::select! {
-                _ = token.cancelled() => Err("MCP operation interrupted by user.".to_string()),
-                timed = send_operation => {
-                    match timed {
-                        Ok(result) => result,
-                        Err(_) => Err("Timed out sending MCP sampling response.".to_string()),
-                    }
-                }
-            }
-        } else {
-            match send_operation.await {
-                Ok(result) => result,
-                Err(_) => Err("Timed out sending MCP sampling response.".to_string()),
-            }
-        };
-        if let Err(error) = &send_result {
-            debug!(
-                server_id = %request.server_id,
-                request_id = ?request.request.id,
-                error = %error,
-                elapsed_ms = start.elapsed().as_millis(),
-                "MCP sampling response send failed"
-            );
-        }
-
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request.request.id,
-            elapsed_ms = start.elapsed().as_millis(),
-            "MCP sampling response send complete"
-        );
-
-        let session_id = context.session_id.clone();
-        let error = send_result.err();
-        app.update(|app| {
-            app.mcp
-                .update_server_request_session(&context.server_id, session_id, error);
-        })
-        .await;
-
-        debug!(
-            server_id = %request.server_id,
-            request_id = ?request.request.id,
-            elapsed_ms = start.elapsed().as_millis(),
-            "Finished MCP sampling flow"
-        );
-        dispatcher.dispatch_many([StreamingAction::McpSamplingFinished], ctx);
-    });
-}
-
-fn spawn_mcp_server_error(
-    app: AppHandle,
-    dispatcher: AppActionDispatcher,
-    server_id: String,
-    request_id: rust_mcp_schema::RequestId,
-    error: rust_mcp_schema::RpcError,
-) {
-    tokio::spawn(async move {
-        let term_size = app.read(|app| app.ui.last_term_size).await;
-        let ctx = AppActionContext {
-            term_width: term_size.width,
-            term_height: term_size.height,
-        };
-
-        let mut context = match app
-            .read(|app| app.mcp.server_request_context(&server_id))
-            .await
-        {
-            Some(context) => context,
-            None => {
-                dispatcher.dispatch_many([StreamingAction::McpSamplingFinished], ctx);
-                return;
-            }
-        };
-
-        let send_result =
-            crate::mcp::client::send_client_error(&mut context, request_id, error).await;
-        let session_id = context.session_id.clone();
-        let error = send_result.err();
-        app.update(|app| {
-            app.mcp
-                .update_server_request_session(&context.server_id, session_id, error);
-        })
-        .await;
-
-        dispatcher.dispatch_many([StreamingAction::McpSamplingFinished], ctx);
-    });
-}
-
-fn spawn_mcp_refresh(app: AppHandle, dispatcher: AppActionDispatcher, server_id: String) {
-    tokio::spawn(async move {
-        let mcp_disabled = app.read(|app| app.session.mcp_disabled).await;
-        if mcp_disabled {
-            app.update(|app| {
-                app.conversation().add_app_message(
-                    AppMessageKind::Info,
-                    "MCP: **disabled for this session**\nMCP refresh skipped.".to_string(),
-                );
-                app.ui.focus_transcript();
-                app.clear_status();
-                app.ui
-                    .end_activity(crate::core::app::ActivityKind::McpRefresh);
-            })
-            .await;
-            dispatcher
-                .dispatch_input_many([StatusAction::ClearStatus], AppActionContext::default());
-            return;
-        }
-
-        let keyring_enabled = app
-            .read(|app| !cfg!(test) && !app.session.startup_env_only)
-            .await;
-        let token_store = McpTokenStore::new_with_keyring(keyring_enabled);
-
-        let mut mcp = app.read(|app| app.mcp.clone()).await;
-        mcp.connect_server(&server_id, &token_store).await;
-        mcp.refresh_tools(&server_id).await;
-        mcp.refresh_resources(&server_id).await;
-        mcp.refresh_resource_templates(&server_id).await;
-        mcp.refresh_prompts(&server_id).await;
-
-        let output = mcp.server(&server_id).map(|server| {
-            crate::commands::build_mcp_server_output(server, keyring_enabled, &token_store)
-        });
-
-        app.update(|app| {
-            app.mcp = mcp;
-            if let Some(message) = output {
-                app.conversation()
-                    .add_app_message(AppMessageKind::Info, message);
-                app.ui.focus_transcript();
-                let term_size = app.ui.last_term_size;
-                if term_size.width > 0 && term_size.height > 0 {
-                    let input_area_height = app.input_area_height(term_size.width);
-                    let mut conversation = app.conversation();
-                    let available_height = conversation
-                        .calculate_available_height(term_size.height, input_area_height);
-                    conversation.update_scroll_position(available_height, term_size.width);
-                }
-            } else {
-                app.conversation()
-                    .set_status(format!("Unknown MCP server: {}", server_id));
-            }
-            app.clear_status();
-            app.ui
-                .end_activity(crate::core::app::ActivityKind::McpRefresh);
-        })
-        .await;
-
-        dispatcher.dispatch_input_many([StatusAction::ClearStatus], AppActionContext::default());
-    });
 }
 
 async fn is_exit_requested(app: &AppHandle) -> bool {
@@ -1135,29 +488,28 @@ async fn drain_action_queue(
                 spawn_model_picker_loader(dispatcher.clone(), request);
             }
             AppCommand::RunMcpTool(request) => {
-                spawn_mcp_tool_call(app.clone(), dispatcher.clone(), request);
+                let context = ExecutorContext::from_app(app.clone(), dispatcher.clone()).await;
+                spawn_mcp_tool_call(context, request);
             }
             AppCommand::RunMcpPrompt(request) => {
-                spawn_mcp_prompt_call(app.clone(), dispatcher.clone(), request);
+                let context = ExecutorContext::from_app(app.clone(), dispatcher.clone()).await;
+                spawn_mcp_prompt_call(context, request);
             }
             AppCommand::RunMcpSampling(request) => {
-                spawn_mcp_sampling_call(app.clone(), dispatcher.clone(), *request);
+                let context = ExecutorContext::from_app(app.clone(), dispatcher.clone()).await;
+                spawn_mcp_sampling_call(context, *request);
             }
             AppCommand::SendMcpServerError {
                 server_id,
                 request_id,
                 error,
             } => {
-                spawn_mcp_server_error(
-                    app.clone(),
-                    dispatcher.clone(),
-                    server_id,
-                    request_id,
-                    error,
-                );
+                let context = ExecutorContext::from_app(app.clone(), dispatcher.clone()).await;
+                spawn_mcp_server_error(context, server_id, request_id, error);
             }
             AppCommand::RefreshMcp { server_id } => {
-                spawn_mcp_refresh(app.clone(), dispatcher.clone(), server_id);
+                let context = ExecutorContext::from_app(app.clone(), dispatcher.clone()).await;
+                spawn_mcp_refresh(context, server_id);
             }
         }
     }
@@ -1747,86 +1099,6 @@ mod tests {
             .await;
         assert_eq!(text, "dumpz");
         assert!(focus_is_input);
-    }
-
-    #[tokio::test]
-    async fn mcp_initializer_completes_when_server_connections_fail() {
-        let app = new_app_handle();
-        let failing_servers = vec![
-            crate::core::config::data::McpServerConfig {
-                id: "alpha".to_string(),
-                display_name: "Alpha".to_string(),
-                base_url: None,
-                command: Some("/definitely-missing-command".to_string()),
-                args: None,
-                env: None,
-                transport: Some("stdio".to_string()),
-                allowed_tools: None,
-                protocol_version: None,
-                enabled: Some(true),
-                tool_payloads: None,
-                tool_payload_window: None,
-                yolo: None,
-            },
-            crate::core::config::data::McpServerConfig {
-                id: "beta".to_string(),
-                display_name: "Beta".to_string(),
-                base_url: None,
-                command: Some("/definitely-missing-command-2".to_string()),
-                args: None,
-                env: None,
-                transport: Some("stdio".to_string()),
-                allowed_tools: None,
-                protocol_version: None,
-                enabled: Some(true),
-                tool_payloads: None,
-                tool_payload_window: None,
-                yolo: None,
-            },
-        ];
-
-        app.update(|app| {
-            app.config.mcp_servers = failing_servers.clone();
-            app.mcp = crate::mcp::client::McpClientManager::from_config(&app.config);
-            app.session.mcp_init_in_progress = true;
-            app.session.mcp_init_complete = false;
-        })
-        .await;
-
-        let (action_tx, mut action_rx) = mpsc::unbounded_channel();
-        let dispatcher = AppActionDispatcher::new(action_tx);
-        let (request_tx, _request_rx) = mpsc::unbounded_channel();
-
-        spawn_mcp_initializer(app.clone(), dispatcher, request_tx);
-
-        let action = tokio::time::timeout(Duration::from_secs(5), action_rx.recv())
-            .await
-            .expect("initializer should dispatch completion")
-            .expect("action should be present");
-        assert!(matches!(
-            action.action,
-            AppAction::Streaming(StreamingAction::McpInitCompleted)
-        ));
-
-        let (init_complete, init_in_progress, alpha_error, beta_error) = app
-            .read(|app| {
-                (
-                    app.session.mcp_init_complete,
-                    app.session.mcp_init_in_progress,
-                    app.mcp
-                        .server("alpha")
-                        .and_then(|server| server.last_error.clone()),
-                    app.mcp
-                        .server("beta")
-                        .and_then(|server| server.last_error.clone()),
-                )
-            })
-            .await;
-
-        assert!(init_complete);
-        assert!(!init_in_progress);
-        assert!(alpha_error.is_some());
-        assert!(beta_error.is_some());
     }
 
     #[test]
